Add createReplEval function
